restructre modal directories and finished add answer
